nvif: split out device control interface definitions
